backend/requirements.txt
# Core LangGraph and LangChain
langgraph>=0.2.40
langchain>=0.3.0
langchain-community>=0.3.0
langchain-openai>=0.2.0
langchain-chroma>=0.1.4

# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0

# Vector Database and Embeddings
chromadb>=0.4.15
sentence-transformers>=2.2.2

# Data Processing
pandas>=2.1.0
numpy>=1.24.0
python-multipart>=0.0.6

# Environment and Configuration
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Utilities
httpx>=0.25.0
aiofiles>=23.2.1
python-json-logger>=2.0.7

# Development
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.9.0
isort>=5.12.0 